requirements.txt
asgiref==3.6.0
bob-builder==0.0.19
boto==2.49.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-cors-middleware==1.5.0
django-heroku==0.3.1
docopt==0.6.2
filelock==3.9.1
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
platformdirs==3.1.1
psycopg2==2.9.5
pypng==0.20220715.0
python-decouple==3.8
qrcode==7.4.2
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
virtualenv-clone==0.5.7
whitenoise==6.4.0
